requirements.txt
DateTime==4.7
greenlet==2.0.0
numpy==1.23.4
pandas==1.5.1
pathlib==1.0.1
pip==22.3
python-dateutil==2.8.2
pytz==2022.6
setuptools==63.2.0
six==1.16.0
SQLAlchemy==1.4.42
zope.interface==5.5.1
apache-airflow-providers-postgres
apache-airflow